+++ argilla-server/tests/unit/api/handlers/v1/webhooks/test_list_webhooks.py
@@ -16,7 +16,7 @@ import pytest
 
 from httpx import AsyncClient
 
-from argilla_server.api.webhooks.v1.enums import WebhookEvent
+from argilla_server.webhooks.v1.enums import WebhookEvent
 from argilla_server.constants import API_KEY_HEADER_NAME
 
 from tests.factories import AdminFactory, AnnotatorFactory, WebhookFactory
