+++ argilla-server/tests/unit/api/handlers/v1/webhooks/test_update_webhook.py
@@ -18,7 +18,7 @@ from uuid import UUID, uuid4
 from httpx import AsyncClient
 from typing import Any
 
-from argilla_server.api.webhooks.v1.enums import WebhookEvent
+from argilla_server.webhooks.v1.enums import WebhookEvent
 from argilla_server.constants import API_KEY_HEADER_NAME
 
 from tests.factories import AdminFactory, AnnotatorFactory, WebhookFactory
